Rename the switches in the CLI
The CLI didn't need as many flags as it had; I doubt people are
typically going to be asking for the description in particular.  As
well, I don't believe "op" and "pre-op" are necessarily what people are
wanting to know, but metadata and PDO info are interesting.  So rename
everything, yay.